requirements.txt
langchain
langchain-community
python-dotenv
streamlit
streamlit-extras
PyPDF2
faiss-cpu==1.7.4
openai
tiktoken
